feat: Refactor plugin initialization and vimrc
This commit refactors the plugin initialization in 'plugins.lua' by
encapsulating the packer bootstrap logic into a function named
'ensure_packer'. This function checks if packer.nvim is installed and
if not, it installs it.

In 'vimrc', the way of requiring Lua modules 'plugins' and 'myconf' is
changed. Instead of using 'lua require(...)', it now uses a 'lua'
heredoc to require these modules. This change makes the code cleaner
and more readable.