feat: Refactor LSP and plugin configurations
This commit includes several changes to the LSP and plugin configurations.
In `lsp_list.lua`, some analyses options are removed for simplicity. In
`myconf.lua`, the keymap for diagnostics is simplified.

In `plugins.lua`, several plugins are added including `nvim-possession`,
`fzf-lua`, and `nvim-lspfuzzy`. The `nvim-bqf` plugin is removed. These
changes aim to improve the overall development experience in Neovim.